moved functions located in utilities that were only used by hicFindEnrichedContacts
